Initial version of smol-cli.
Initial functionality of the smol-cli mimics the smolpub.sh bash script.
You supply a path to a post and it will create or update the post
depending on whether it already exists. The script does some very basic
error checking at the moment. The post format mirrors the one used by
smolpub.sh.